+++ visualization/app/codeCharta/ui/codeMap/codeMap.component.ts
@@ -3,7 +3,6 @@ import { Component, Inject, AfterViewInit, ElementRef, OnDestroy } from "@angula
 import { Store } from "../../state/angular-redux/store"
 import { isLoadingFileSelector } from "../../state/store/appSettings/isLoadingFile/isLoadingFile.selector"
 import { ThreeViewerService } from "./threeViewer/threeViewer.service"
-import { CodeMapMouseEventServiceToken } from "../../services/ajs-upgraded-providers"
 import { sharpnessModeSelector } from "../../state/store/appSettings/sharpnessMode/sharpnessMode.selector"
 import { CodeMapMouseEventService } from "./codeMap.mouseEvent.service"
 import { skip, tap } from "rxjs"
@@ -30,7 +29,7 @@ export class CodeMapComponent implements AfterViewInit, OnDestroy {
 		@Inject(IsAttributeSideBarVisibleService) public isAttributeSideBarVisibleService: IsAttributeSideBarVisibleService,
 		@Inject(Store) private store: Store,
 		@Inject(ThreeViewerService) private threeViewerService: ThreeViewerService,
-		@Inject(CodeMapMouseEventServiceToken) private codeMapMouseEventService: CodeMapMouseEventService,
+		@Inject(CodeMapMouseEventService) private codeMapMouseEventService: CodeMapMouseEventService,
 		@Inject(ElementRef) private elementReference: ElementRef
 	) {}